+++ src/devicecontroller/DefaultDeviceController.ts
@@ -1561,16 +1561,6 @@ export default class DefaultDeviceController
       trackConstraints.frameRate = trackConstraints.frameRate || {
         ideal: this.videoInputQualitySettings.videoFrameRate,
       };
-      // TODO: try to replace hard-code value related to videos into quality-level presets
-      // The following configs relaxes CPU overuse detection threshold to offer better encoding quality
-      // @ts-ignore
-      trackConstraints.googCpuOveruseDetection = true;
-      // @ts-ignore
-      trackConstraints.googCpuOveruseEncodeUsage = true;
-      // @ts-ignore
-      trackConstraints.googCpuOveruseThreshold = 85;
-      // @ts-ignore
-      trackConstraints.googCpuUnderuseThreshold = 55;
     }
     if (kind === 'audio' && this.supportSampleRateConstraint()) {
       trackConstraints.sampleRate = { ideal: DefaultDeviceController.defaultSampleRate };